iou_metric

+++ iou.py
@@ -28,7 +28,7 @@ def mean_iou(imgs_true, imgs_pred):
             scores[i] = 1
         else:
             # scores[i] = (iou_thresholds <= iou(imgs_true[i], imgs_pred[i])).mean()
-            scores[i] = np.mean([iou(imgs_true[i], threshold<=imgs_pred[i]) for threshold in iou_thresholds])
+            scores[i] = np.mean([iou(imgs_true[i], (threshold<=imgs_pred[i]).astype('uint8')) for threshold in iou_thresholds])
             
     return scores.mean()
